fixed-Slice1-review
