--- src/Command.hpp
@@ -1,18 +1,0 @@
-/**
- * @file
- * @brief     Command class
- * @author    Quinten Perquin
- * @license   See LICENSE
- */
-
-#ifndef COMMAND_HPP
-#define COMMAND_HPP
-#include "Vector2f.hpp"
-
-class Command
-{
-public:
-    Command(){}
-};
-
-#endif
