--- src/MPU9250_Interface.hpp
@@ -1,17 +1,0 @@
-/**
- * @file
- * @brief     MPU9250Interface Class
- * @author    Quinten Perquin
- * @license   See LICENSE
- */
-
-#ifndef MPU9250_INTERFACE_HPP
-#define MPU9250_INTERFACE_HPP
-
-class MPU9250Interface
-{
-public:
-    MPU9250Interface(){}
-};
-
-#endif
