--- src/Navigation_Engine.hpp
@@ -1,25 +1,0 @@
-/**
- * @file
- * @brief     NavigationEngine Class
- * @author    Quinten Perquin
- * @license   See LICENSE
- */
-
-#ifndef NAVIGATION_ENGINE_HPP
-#define NAVIGATION_ENGINE_HPP
-
-#include "IO_Interface_Listener.hpp"
-#include "Position_Module.hpp"
-
-class NavigationEngine : public IOInterfaceListener {
-  public:
-    NavigationEngine() {
-    }
-
-  private:
-    void onReceive(const Command &command) {
-    }
-    PositionModule positionModule;
-};
-
-#endif
