+++ src/command.hpp
@@ -1,0 +1,26 @@
+/**
+ * @file
+ * @brief     Command class
+ * @author    Quinten Perquin
+ * @license   See LICENSE
+ */
+
+#ifndef COMMAND_HPP
+#define COMMAND_HPP
+
+#include "vector2f.hpp"
+
+class Command {
+  public:
+    /**
+     * @brief creates a Command object
+     *
+     * Constructor for a Command object without arguments.
+     *
+     * @return a Command object
+     */
+    Command() {
+    }
+};
+
+#endif
