+++ src/mpu9250_interface.hpp
@@ -1,0 +1,24 @@
+/**
+ * @file
+ * @brief     MPU9250Interface Class
+ * @author    Quinten Perquin
+ * @license   See LICENSE
+ */
+
+#ifndef MPU9250_INTERFACE_HPP
+#define MPU9250_INTERFACE_HPP
+
+class MPU9250Interface {
+  public:
+    /**
+     * @brief creates a MPU9250Interface object
+     *
+     * Constructor for a MPU9250Interface object without arguments.
+     *
+     * @return a MPU9250 object
+     */
+    MPU9250Interface() {
+    }
+};
+
+#endif
